YARR_Server/requirements.txt
python-socketio[client]
python-socketio[asyncio_client]
aiomysql
aiohttp

#mybe still needed bu not sure:
python-dotenv
mysql-connector
mysql-connector-python
dnspython
